get rid of Shrinker.create_lit which is subsumed by minimal_term
One difference is that Shrinker.create_lit would try to randomly
generate literal values (calling small_int.gen for example to create
an integer literal), while minimal_term only uses a fixed value. We
believe that randomly generating literals is not useful for
shrinking. Either the value does not matter at all, and then using
a fixed literal is just fine (and it makes it easier to recognize in
the output that this part of the term does not matter); or the value
does matter to reproduce the bug, and trying to generate the "right"
values randomgly is not going to work anyway.