patterns variables should be unique, but can shadow the environment
The previous implementation would guarantee that the variable picked
for each pattern would be distinct from any variable in the
environment. This is more strict than necessary; we now keep
a seperate set of variables of the current pattern.

Note: this commit also fixes a bug where

    | other ->
      ...
      env := add_var var typ !env;

was used instead of

    | other ->
      ...
      env := add_var var other !env;

with the result of asigning the wrong type to a pattern variable
whenever it was under a tuple construct. (It seems that this didn't
result in a test failure because we currently do not type-check
patterns very carefully).